+++ serializers/project.js
@@ -2,15 +2,8 @@ const JSONAPISerializer = require('jsonapi-serializer').Serializer;
 
 function Project(data) {
   return new JSONAPISerializer('projects', {
-    attributes: ['name', 'defaultEnvironment'],
-    defaultEnvironment: {
-      ref: 'id',
-      included: false,
-    },
-    typeForAttribute: (type) => {
-      if (type === 'defaultEnvironment') { return 'environments'; }
-      return type;
-    },
+    attributes: ['name'],
+    keyForAttribute: 'underscore_case',
   }).serialize(data);
 }