+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ### Changed
 - Commands - Drop environment variables.
+- Commands - Remove command run and install.
+- Command Generate - Install Forest Admin middleware.
 
 ## RELEASE 2.6.1 - 2019-10-30
 ### Fixed